Move DomainCoreApi from sp-domains to domain-runtime-primitives
